refactoring of whitelist services / split into common lib for rules and whitelist handling services to avoid ciruclar imports
