+++ flowapp/services/base.py
@@ -1,6 +1,6 @@
-from flowapp import messages
-from flowapp.constants import ANNOUNCE
-from flowapp.models import RTBH
+from flowapp import db, messages
+from flowapp.constants import ANNOUNCE, RuleOrigin, RuleTypes
+from flowapp.models import RTBH, RuleWhitelistCache
 from flowapp.output import Route, RouteSources, announce_route
 
 
@@ -16,3 +16,12 @@ def announce_rtbh_route(model: RTBH, author: str) -> None:
             command=command,
         )
         announce_route(route)
+
+
+def add_rtbh_rule_to_cache(model: RTBH, whitelist_id: int, rule_origin: RuleOrigin = RuleOrigin.USER) -> None:
+    """
+    Add RTBH rule to whitelist cache
+    """
+    cache = RuleWhitelistCache(rid=model.id, rtype=RuleTypes.RTBH, whitelist_id=whitelist_id, rorigin=rule_origin)
+    db.session.add(cache)
+    db.session.commit()
